Change API to have state and setState in the same function
It seems to make more sense to have both state and setState available in
the mapping function, as the state mutation commonly depends on the
state itself.

It might be clever to abstract this to a reducer-action kind of state
handling (like Redux), but this seems like a good enough split for now.